[REFACTOR] Rename folder from src to core
Renamed the project’s main source directory from `src` to `core` to better reflect its role as the central codebase. Updated imports and references accordingly.
